refactor: Split out non-method based methods to a base function interface
